Enable shared code between /web and /server
Until now, we've been able to share types from /common to both our
/server and /web codebases. However, importing any code into /server
from /common that wasn't stripped out by TypeScript (so, anything but
types), caused our directory structure to change. It broke the build
in a non-obvious way, and has been a frustrating limitation.

I've slowly and surely been upgrading our Webpack / TypeScript
dependencies, which was part of the puzzle.

This commit turns our /common directory into a proper standalone module,
with a single generated index.ts file, a build step, and type
annotations.

For consistency, we now rely on [Yarn workspaces](https://classic.yarnpkg.com/en/docs/workspaces/) on both the /web and /server packages. Treating /common as a standalone module has encapsulation benefits, and will let us manage a separate dependency list, build steps, etc. for /common.

In the process of doing this work I realized that we have a lot of (near) options for solving this task. Here are a few off the top of my head:

- Yarn Workspaces
- Yarn / NPM Private Packages
- Yarn link
- "file: ../" dependencies
- TypeScript Project References
- TypeScript Internal Modules
- TypeScript Path Aliases
- TypeScript Path Mapping (different thing)
- Lerna
- Oh… have I not even mentioned webpack yet?

We already using some of these, which made the task frustrating. The goal is to pick a subset that works for us, and strictly stick to that subset.

I've already cleared a _huge_ list of TODOs, FIXMEs, and STOPSHIPS, but
here are the last few I'll get to tomorrow:

- Fixme: Remove all `@ts-ignore` lines.
- Todo: ensure bundle sizes haven't exploded.
- Todo: Ensure nothing has gotten in the way of our IDE-ish TypeScript capabilities.

This commit also adds some trivial `real code` imports into /server,
so we can test it out right away.

For quick checks, I found the following adequate:

`cd server && tsc -b && cd ../web && yarn build`

or,

`cd server && yarn start` (tests whether node / commonjs imports are
finding /common imports)

or,

`cd common && tsc`

To run a final test of this change, I wanted to be sure that things would work
from scratch. The test plan below is a bit overkill, and it will take a while,
but it was worth the peace of mind.

```
docker ps -a
docker images
rm -rf web/node_modules/ server/node_modules/ node_modules/ web/dist/ server/js common/js

docker-compose up
```

After waiting for a long time, I went to localhost:9000 and… it worked!  🎉

The current monorepo structure was initially introduced in
https://github.com/mozilla/voice-web/commit/edcf9ae56948b517ce505c6af7a975f0790b3db7